Fix horizontal scrollbar in phone orientation
If the nav was being shown full-width and there were page titles longer
than the screen width, it wasn't truncating/ellipsifying them, which
resulted in the page getting a horizontal scrollbar (which messed up
all the boxes' the background colors, since they filled the screen
width and not beyond). Fixed by making sure it truncates.